+++ packages/react/Title/CHANGELOG.md
@@ -2,7 +2,19 @@
 
 Tilte is a basic component that allows you to give titles and more hierarchy to blocks of text for sections or header components
 
-## 2022-09-01
+## 2022-11-02 `1.4.0`
+
+### 🎉 New features
+
+- Added displayName to component `Title.Skeleton`. ([#52](https://github.com/TiendaNube/nimbus-design-system/pull/52) by [@juniorconquista](https://github.com/juniorconquista))
+
+## 2022-09-12 `1.3.0`
+
+### 🎉 New features
+
+- Added `textAlign` properties to the Component. ([#23](https://github.com/TiendaNube/nimbus-design-system/pull/23) by [@juniorconquista](https://github.com/juniorconquista))
+
+## 2022-09-01 `1.2.0`
 
 ### 📚 3rd party library updates
 
@@ -12,7 +24,7 @@ Tilte is a basic component that allows you to give titles and more hierarchy to 
 
 - Added `color` properties to the Component Title. ([#21](https://github.com/TiendaNube/nimbus-design-system/pull/21) by [@juniorconquista](https://github.com/juniorconquista))
 
-## 2022-08-22
+## 2022-08-22 `1.1.1`
 
 ### 📚 3rd party library updates
 
@@ -26,13 +38,13 @@ Tilte is a basic component that allows you to give titles and more hierarchy to 
 - Added `as`, `width` and `height` properties to the Component `Title.Skeleton`. ([#15](https://github.com/TiendaNube/nimbus-design-system/pull/15) by [@juniorconquista](https://github.com/juniorconquista))
 - Added stories on Component `Title.Skeleton`. ([#15](https://github.com/TiendaNube/nimbus-design-system/pull/15) by [@juniorconquista](https://github.com/juniorconquista))
 
-## 2022-08-21
+## 2022-08-21 `1.1.0`
 
 ### 💡 Others
 
 - Added new unit tests to ensure text align styles are applied. ([#13](https://github.com/TiendaNube/nimbus-design-system/pull/13) by [@juniorconquista](https://github.com/juniorconquista))
 
-## 2022-08-20
+## 2022-08-20 `1.1.0`
 
 ### 📚 3rd party library updates
 
@@ -46,7 +58,7 @@ Tilte is a basic component that allows you to give titles and more hierarchy to 
 - Added `webpack@5.74.0`. ([#11](https://github.com/TiendaNube/nimbus-design-system/pull/11) by [@juniorconquista](https://github.com/juniorconquista))
 - Added `webpack-cli@4.10.0`. ([#11](https://github.com/TiendaNube/nimbus-design-system/pull/11) by [@juniorconquista](https://github.com/juniorconquista))
 
-## 2022-08-09
+## 2022-08-09 `1.0.2`
 
 <!-- ### 📚 3rd party library updates -->